Updated git submodules
Project: why3 master 7398072a672b9a58897a4e447610e16d0609dbdc

S911-045 Fix previous commit that added potentially empty warnings

The commit 9a17d3b3e07130 always add an extra warning (per call to
gnatwhy3) which may be empty (when no debug flag are set: always). This
was invisible in mailservers because of the filtering of empty lines that
are done by the testsuite.
This commit checks if the string used for debug is empty before reporting
it as a warning.

* src/gnat/gnat_main.ml
(ending): Do not add a warning when the message is empty

Change-Id: I2fbabe69fea0f2ab7b67852724a035d817a60e6f